feat(display): initial rough outline for central "display" module
This module is intended to handle all transformations of the original
raw bytes, into an encoded, cached, duplicate of UTF-8 text to display
to the user.
It serves as a kind of "utility" module of the core, that should be used
to handle a lot of common logic across front-ends.